requirements.txt
numpy
pandas
tables
grpcio-tools
redis
python-dotenv
mypy-protobuf